fix the same flaky test (hopefully final fix)
Saw a _different_ break in CI in the same test as the previous commit, due to overflow https://github.com/Layr-Labs/eigenlayer-contracts/actions/runs/6041156370/job/16393490571 -- this failure will only occur with a super high share amount _and_ a duplicate array entry. by limiting the fuzzed input size we fix the issue.

+++ src/test/unit/DelegationUnit.t.sol
@@ -1018,7 +1018,7 @@ contract DelegationUnitTests is EigenLayerTestHelper {
      * who the `staker` is delegated to has in the strategies
      * @dev Checks that there is no change if the staker is not delegated
      */
-    function testDecreaseDelegatedShares(address staker, IStrategy[] memory strategies, uint256 shares, bool delegateFromStakerToOperator) public {
+    function testDecreaseDelegatedShares(address staker, IStrategy[] memory strategies, uint128 shares, bool delegateFromStakerToOperator) public {
         // sanity-filtering on fuzzed input length
         cheats.assume(strategies.length <= 64);
         // register *this contract* as an operator
